detect_rules/kql/requirements.txt
azure-identity==1.24.0
azure-mgmt-loganalytics==14.0.0
azure-mgmt-resource==24.0.0
pytest==8.4.2
pytest-cov==6.3.0
pyyaml==6.0.2
prometheus-client==0.22.1
cachetools==6.2.0
flake8==7.3.0